@@ -11,32 +11,37 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.1.0",
-    "@angular/common": "^17.1.0",
-    "@angular/compiler": "^17.1.0",
-    "@angular/core": "^17.1.0",
-    "@angular/forms": "^17.1.0",
-    "@angular/platform-browser": "^17.1.0",
-    "@angular/platform-browser-dynamic": "^17.1.0",
-    "@angular/platform-server": "^17.1.0",
-    "@angular/router": "^17.1.0",
-    "@angular/ssr": "^17.1.0",
+    "@angular/animations": "^17.2.2",
+    "@angular/cdk": "^17.2.1",
+    "@angular/common": "^17.2.2",
+    "@angular/compiler": "^17.2.2",
+    "@angular/core": "^17.2.2",
+    "@angular/forms": "^17.2.2",
+    "@angular/material": "^17.2.1",
+    "@angular/platform-browser": "^17.2.2",
+    "@angular/platform-browser-dynamic": "^17.2.2",
+    "@angular/platform-server": "^17.2.2",
+    "@angular/router": "^17.2.2",
+    "@angular/ssr": "^17.2.1",
     "@ng-bootstrap/ng-bootstrap": "^16.0.0",
     "@popperjs/core": "^2.11.8",
+    "@types/google-one-tap": "^1.2.6",
     "bootstrap": "^5.3.2",
     "express": "^4.18.2",
+    "jwt-decode": "^4.0.0",
     "localstorage-polyfill": "^1.0.1",
     "ng-bootstrap": "^1.6.3",
-    "ngx-bootstrap": "^10.3.0",
+    "ngx-bootstrap": "^12.0.0",
+    "ngx-clipboard": "^16.0.0",
     "rxjs": "~7.8.0",
     "tslib": "^2.3.0",
     "zone.js": "~0.14.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.1.0",
-    "@angular/cli": "^17.1.0",
-    "@angular/compiler-cli": "^17.1.0",
-    "@angular/localize": "^17.1.0",
+    "@angular-devkit/build-angular": "^17.2.1",
+    "@angular/cli": "^17.2.1",
+    "@angular/compiler-cli": "^17.2.2",
+    "@angular/localize": "^17.2.2",
     "@types/express": "^4.17.17",
     "@types/jasmine": "~5.1.0",
     "@types/node": "^18.18.0",
