@@ -22,6 +22,7 @@ import { BlogComponent } from './blog/blog.component';
 import { SupportComponent } from './support/support.component';
 import { GetStartedComponent } from './get-started/get-started.component';
 import { FormsModule } from '@angular/forms';
+import { WorkspaceModule } from './workspace/workspace.module';
 
 @NgModule({
   declarations: [
@@ -38,18 +39,16 @@ import { FormsModule } from '@angular/forms';
     SurveyStatisticsComponent,
     BlogComponent,
     SupportComponent,
-    GetStartedComponent
-
-    
+    GetStartedComponent,
   ],
 
-  imports: [BrowserModule, AppRoutingModule, NgbModule, SharedModule,HttpClientModule,FormsModule],
+  imports: [BrowserModule, AppRoutingModule, NgbModule, SharedModule,HttpClientModule,FormsModule,WorkspaceModule],
   // providers: [
   //   provideClientHydration()
   // ],
 
   providers: [
-    { provide: HTTP_INTERCEPTORS, useClass: JwtInterceptor, multi: true },
+    { provide: HTTP_INTERCEPTORS, useClass: JwtInterceptor, multi: true }
   ],
   bootstrap: [AppComponent],
 })
