@@ -1,17 +1,21 @@
 ﻿using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Online_Survey.DTOs;
 using Online_Survey.Models;
+using Online_Survey.PocoClass;
+using System.Linq;
 
 namespace Online_Survey.Data
 {
     public class UserRepository : IUserRepository
     {
         private readonly InternshipOnlineSurveyContext _ef;
-
-        public UserRepository(IConfiguration config)
+        private InternshipOnlineSurveyContext _ef1;
+        public UserRepository(IConfiguration config, InternshipOnlineSurveyContext ef1)
         {
             _ef = new InternshipOnlineSurveyContext(config);
+            _ef1 = ef1;
         }
 
         bool IUserRepository.SaveChange()
@@ -25,6 +29,11 @@ namespace Online_Survey.Data
             {
                 _ef.Add(data);
             }
+        }
+
+        IQueryable<QuestionTable> IUserRepository.QuestionOption()
+        {
+            return _ef.QuestionTables.Include(q => q.OptionTables);
         }
     }
 }
