@@ -6,14 +6,14 @@ namespace Online_Survey.DTO
     {
         public string SurveyorId { get; set; }
 
+        public string Description { get; set; }
+
         public DateOnly DateCreated { get; set; }
 
-        public DateOnly LaunchDate { get; set; }
+        public DateOnly? LaunchDate { get; set; }
 
-        public DateOnly EndDate { get; set; }
+        public DateOnly? EndDate { get; set; }
 
-        public TimeOnly StartTime { get; set; }
-
-        public int EndTime { get; set; }
+        public TimeOnly? StartTime { get; set; }
     }
 }
