@@ -1,5 +1,7 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, Inject, OnInit } from '@angular/core';
 import { AccountService } from './account/account.service';
+import { DOCUMENT } from '@angular/common';
+import { GlobalserviceService } from '../globalservice/globalservice.service';
 
 @Component({
   selector: 'app-root',
@@ -9,13 +11,13 @@ import { AccountService } from './account/account.service';
 export class AppComponent implements OnInit {
 
 
-  constructor(private accountService : AccountService){}
+  constructor(private accountService : AccountService,@Inject(DOCUMENT) private document: any,private globalservice : GlobalserviceService){}
 
 
   ngOnInit(): void {
 
     this.refreshUser()
-   
+    this.globalservice.FrontendUrl = this.document.location.href;
   }
 
   private  refreshUser(){
