@@ -6,6 +6,7 @@ import { CompanyComponent } from './company/company.component';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { DepartmentComponent } from './department/department.component';
 import { SurveyerWorkspaceComponent } from './surveyer-workspace/surveyer-workspace.component';
+import { SidebarComponent } from './sidebar/sidebar.component';
 
 
 
@@ -15,6 +16,7 @@ import { SurveyerWorkspaceComponent } from './surveyer-workspace/surveyer-worksp
     CompanyComponent,
     DepartmentComponent,
     SurveyerWorkspaceComponent,
+    SidebarComponent,
   
    
   ],
@@ -24,6 +26,7 @@ import { SurveyerWorkspaceComponent } from './surveyer-workspace/surveyer-worksp
     ReactiveFormsModule,
     FormsModule
     
-  ]
+  ],
+  exports: [SidebarComponent],
 })
 export class WorkspaceModule { }
